Reuse first focusable element logic to focus an element on the configure git screen
